feature: support for tracking in Lettuce versions 6.5 and above
