+++ CHANGES.md
@@ -14,6 +14,7 @@ Release Notes.
 * Fix retransform failure when enhancing both parent and child classes.
 * Add support for `dameng(DM)` JDBC url format in `URLParser`.
 * Fix RabbitMQ Consumer could not receive handleCancelOk callback.
+* Support for tracking in lettuce versions 6.5.x and above.
 
 All issues and pull requests are [here](https://github.com/apache/skywalking/milestone/236?closed=1)
