+++ apm-sniffer/apm-sdk-plugin/lettuce-plugins/lettuce-5.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/lettuce/v5/define/DefaultEndpointInstrumentationV5.java
@@ -1,0 +1,33 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package org.apache.skywalking.apm.plugin.lettuce.v5.define;
+
+import java.util.Collections;
+import java.util.List;
+import org.apache.skywalking.apm.agent.core.plugin.WitnessMethod;
+import org.apache.skywalking.apm.plugin.lettuce.common.define.DefaultEndpointInstrumentation;
+import org.apache.skywalking.apm.plugin.lettuce.v5.constant.Constants;
+
+public class DefaultEndpointInstrumentationV5 extends DefaultEndpointInstrumentation {
+
+    @Override
+    protected List<WitnessMethod> witnessMethods() {
+        return Collections.singletonList(Constants.WITNESS_LETTUCE_5X_METHOD);
+    }
+}
