+++ apm-sniffer/apm-sdk-plugin/pom.xml
@@ -87,7 +87,7 @@
         <module>netty-socketio-plugin</module>
         <module>httpclient-3.x-plugin</module>
         <module>play-2.x-plugin</module>
-        <module>lettuce-5.x-plugin</module>
+        <module>lettuce-plugins</module>
         <module>avro-plugin</module>
         <module>finagle-6.25.x-plugin</module>
         <module>quasar-plugin</module>
